restructure into workspaces with top level integration testing
