+++ crates/client/run.sh
