+++ crates/client/src/config.rs
@@ -1,6 +1,5 @@
 use std::error::Error;
 
-use clap::{arg, command, Parser};
 use serde::Deserialize;
 
 #[derive(Deserialize, Debug)]
@@ -11,11 +10,4 @@ pub struct Config {
 
 pub fn get_config() -> Result<Config, Box<dyn Error>> {
     Ok(envy::from_env::<Config>()?)
-}
-
-#[derive(Parser)]
-#[command(author, version, about, long_about = None)]
-pub struct Args {
-    #[arg(short, long, default_value = "anon")]
-    pub username: String,
 }
