+++ crates/client/src/lib.rs
@@ -1,16 +1,8 @@
-mod config;
-
-use clap::Parser;
-use config::{get_config, Args};
+pub mod config;
 use tokio::net::TcpStream;
 
 #[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let args = Args::parse();
-
-    println!("Hello, {}!", args.username);
-
-    let config_values = get_config()?;
+pub async fn run(config_values: &config::Config) -> Result<(), Box<dyn std::error::Error>> {
     let address = format!("{}:{}", config_values.host, config_values.port);
     let mut stream = TcpStream::connect(&address).await?;
     println!("Connected to the server at {}", address);
