+++ crates/server/Cargo.toml
@@ -3,6 +3,9 @@ name = "server"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+path = "src/lib.rs"
+
 [dependencies]
 tokio = { version = "1", features = ["macros", "net", "rt-multi-thread"] }
 serde = { version = "1", features = ["derive"]}
