+++ crates/server/run.sh
